+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ==========
 
+* Fix template pagination for view search.
 * Fix request.is_ajax in views.
 
 4.0.4 (2026-04-14)